scripts/demo/requirements.txt
streamlit
torch
pytorch-crf
numpy
scikit-learn